api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.33
botocore==1.37.33
botocore-stubs==1.38.13
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.3
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
types-awscrt==0.27.1
types-s3transfer==0.12.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.34.1
